requirements.txt
alembic==1.7.6
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
DateTime==4.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
zope.interface==5.4.0
